cron/requirements.txt
garminconnect>=0.2
psycopg2-binary>=2.9